Added SynchronizeDataNode
This method reads some general data from the given node, then creates a subscription with given eventhandler, then fetches all historical data if relevant. The historical data is fetched in order, oldest first, with a maximum number of results per query.
For each loop, a callback function is called with the results. This solution is due to the fact that we have a continuationVariable, which is efficient to use for this kind of repeated read.

Also cleaned up a bit, and changed some reference rules to make VS behave.