Reverted to using NodeId as identifier in BufferedNode
The UniqueId system was just dumb, it didn't really work as intended, and it still had complete dependency on Opc.Ua, except instead of an explicit dependency, it just depended implicitly on the internals, which is much worse. This is just back to the old system, except of course to fit with the new Pusher pattern UAClient has to be passed to the pusher.

This does seem to almost make Extractor.cs obsolete, however I think we can introduce some proper paralellism by making each extractor manage multiple pairs of pusher/client, in theory at least.