Made pusher fail fast
No point in spamming CDF with pointless requests that are almost certain to fail, much better to just terminate on any error,
as most likely that will mean that most future requests will fail as well, or faulty data will be introduced.

Any better handling than that will be very difficult, and most likely not worth it at all.