Use ExternalId as root assetId.
This should, in theory, allow us to get properties on root nodes, and
make it easier to specify root. The changes to the load are tiny, in
practice it only means a single more read before browsing, and one more
node to map.

As an aside, I also fixed a few small issues