Replace library used for pushing to influx.
The new one is a bit simpler in some ways, but has more features, so it
allows us to test for first/latest timestamp in influx as well. This is
done on all historizing variables.

This commit also adds a new method to fetch properties cleverly, by
keeping track of all pending nodes.