Fix multiple pushers (#50)
* Some fixes, add test.

Some major changes to structure were required to make this work
properly, in particular the call to synchronizenodes had to be moved out
of the pushers, and to make sure that we don't fetch properties multiple
times, these had to be isolated.

* Replace library used for pushing to influx.

The new one is a bit simpler in some ways, but has more features, so it
allows us to test for first/latest timestamp in influx as well. This is
done on all historizing variables.

This commit also adds a new method to fetch properties cleverly, by
keeping track of all pending nodes.