Add support for arrays and strings
Strings are, by default, every other datatype than those already
registered as numeric, even custom datatypes.

Arrays require the ArrayDimensions property to be set, it is then
compared to the MaxArraySize config option to determine whether or not
to push the variable.

ExternalIds for array-type variables are generated by simply adding
[index] to the end. They are then created as individual timeseries under
a "virtual" object representing the array itself.