Fix issue with tests.
The test servers were all writing to the same sql file, this was fine
when we didn't care about the contents of the data (the historical
points were all of the same type anyway), but now that we have array
datapoints on the same tag as double datapoints this caused problems.

Also improved error handling in HistoryRead, notably making it so that
the history read tasks are awaited, and thus thrown properly.