Improve config structure a bit. (#53)
* Improve config structure a bit.

Done in preparation of Event configuration, which will be quite
extensive.

Remove the BulkSizes config object, which now only contained opcua
config, so it might as well be moved into the UAConfig object.

Removed bulk sizes from cdf config, this generally isn't configurable on
cognite products, and there is little reason for it to be, if there are
concerns about speed that is better fixed in a different way.

Added ExtractionConfig object, which now contains all the configuration
concerning what nodes to extract. Most of this was in UAConfig, but it
is better to narrow UAConfig down and move this to a new object.

Changed indentation to make config more readable (four spaces).

* Cleanup test config, make config more modular.

It should be possible to use only a minimal subset of the default
config, anything else should have default values and be capable of
dealing with null values in config files.