Add configuration for events
Automatically map out the type-hierarchy for events and get all
attributes for a given event type.

This approach is slightly limited in that we use browsenames as unique
names, which they are not. It isn't possible to ignore just one property
with a given browsename and include properties with the same name on
subtypes. This should be a small issue, however.