First commit of history read events.
To make things a bit cleaner, separate the continuationPoint/HistoryRead
logic out of the main history read function, as it is mostly common for
events and data.

Add new config options for history read chunksize, enable history read,
and the browseName of the property that gives us a property to compare
time to, to avoid reading all events each time.

Pushing of events to CDF required some changes, as there is a
possibility of overlapping events. There is no clean way to fix this, so
on failed creation of events, we retry without the duplicate ids.