Add tests for events.
Required some changes, and a few modifications were made in order to
fix discovered issues:

We now use "Source" in order to describe the
actual source of the event, which may be a timeseries.

Two different HTTP clients configurations are provided, in order to
allow quicker failure on events and data.

Lots of bugfixes