Feature/events (#61)
Large new feature that adds support for events to the extractor. Events are dynamically mapped out on the server, where we identify properties of events to be pushed to CDF by BrowseName.

Serverside filtering is used to select which events and event-properties we want. The user must specify a list of events (in order for this to work on event-heavy servers), and a list of event emitters. The SourceNode property on each event must also be set, and match a known node in the hierarchy.

Events are filtered clientside as well.

Some CDF specific event properties like type, subtype, starttime, endtime can be set by having a property with each name on the event, otherwise they are either set to null, or constructed from "Time", "EventType", etc.

Tests required cloning cognitedata/python-opcua, which is our fork of python-opcua containing the fixes. Once they are integrated into a release we can go back to using pip.

History read is also implemented. The latest time events were run is saved to a local file and read on startup, in order to avoid overlap, as the same method cannot be used as with timeseries. There will still be overlap, which will cause failures in CDF. This is handled by retrying after removing all conflicting events.

There is a lot of code, the greatest evidence that it works is in the test, which is quite thorough.

Note that python-opcua does not support serverside filtering of events during historyRead. The filter should be the same, however.