Refactor and bugfixing
Fixes a couple significant issues with the last few patches. Refactors
the state into a single object, and adds a buffer to store datapoints
while historyread is running, to prevent loss of points.

This separates the fetching of latest timestamp out of the initial
creation of points, into a separate method on IPusher. In theory this
might take slightly longer for CDF, but in practise I've found that the
initial checking is very quick. This will make the first iteration a
fair bit faster, with every subsequent one slower.