Cleanup in CDFPusher, fixed an issue in EventTests
The issue in eventTests is probably rare, but it was nice to fix.

In CDFPusher I moved the testing of datatype to the initial loop, which
is cleaner and means that we push the full number of datapoints with
each push to CDF (potentially).

I also moved the storage of Historizing and the check of consistent type
to checking against the main extractor state, to do this I had to add a
dictionary to map externalId to NodeId, which still resulted in less
code.