Feature/node discovery (#75)
Two strategies for node discovery:

The first is a simple config option to re-browse every x minutes. Making rebrowse possible required changing UAClient slightly. Now there is both a global "visitedNodes" which is used to decide which nodes should callback, and a local "localVisitedNodes" for each browse loop, which is used to avoid loops in the node tree.

The second is a neat system of using auditEvents in OPC-UA to discover nodes. From what I can tell, few implementations actually use this, but if they do it is far more efficient and clean than a periodic browse.

It works by listening to AuditAddNodesEvent and AuditAddReferencesEvent, which simply provide the NodesToAdd or ReferencesToAdd lists from their respective services on the server to an auditing client, then they use the SourceNode or the ParentNode given in the events to trigger a re-browse from those nodes. If the node-ids are unknown, we do nothing, so this will work even if auditing events are fired for parts of the tree that aren't mapped.

Python-opcua has no native support, but the OPC-UA type hierarchy is re-created, so it was fairly simple to create a system to manually fire a couple of auditing events. The parsing clientside was done by code from the SDK, so the structure should be correct.