Fix to stubborn old bug in subscriptions
This bug has been known since the beginning of development, the issue is
that on a reconnect old subscriptions hang around, and new subscriptions
are added seperately, so we get double subscriptions for all values.

As it turns out, the issue was fairly simple to fix. An earlier attempt
at a patch made it so that we searched the list of created subscriptions
for the previous subscription DisplayName before adding new
subscriptions, but as it turns out, on a restart the SDK changes the
displayname to add a number at the end, for some reason. Once this was
dealt with the issue was easily resolved.