Add FailureBuffer system
Instead of buffering locally on the pushers ,have hte pushers return the
failed points after each push, and a keep a separate "FailureBuffer"
class to handle the actual buffering.

The system as it works now allows for both using an InfluxPusher as a
drain, and a raw binary file as before, however it now also supports
buffering on InfluxPusher and any future pushers with very few changes
to the pushers.

The plan is to add support for events as well, though this will require
influxdb support for events.

This commit also adds a couple tests and improves a couple old tests.