Feature/influxdb events (#87)
Support for events in influxdb, using events.[full normal id] as measurement names, to avoid overlapping with values when the source node is a variable.

The implementation is fairly simple: use string-valued measurements in influxdb, the value of each point is the "Message" from OPC-UA, and all metadata is added as tags. Since the variety of metadata is fixed, this means that we have a limited scope of tags, which should make the results quite readable. If a metadata value is never specified it will generally not appear as a tag.

This is also a nice demonstration of the power of the generic pusher system, adding support for events required no changes to anything other than the influxdb pusher itself, and even if it is completely broken it should not affect anything else.