New common functionality for tests
Adds a new ExtractorTester class to the common test code which is meant
to be a container for the extractor, uaClient, pushers etc. and to help
with making the code more compact and allow for changes to all tests
much more easily.

The tests should largely be the same, except that they may take a little
less time. The change is that they are far more readable.