Add FailureBuffer system (#85)
Instead of buffering locally on the pushers, have the pushers return the failed points after each push, and a keep a separate "FailureBuffer" class to handle the actual buffering.

The system as it works now allows for both using an InfluxPusher as a drain, and a raw binary file as before, however it now also supports buffering on InfluxPusher and any future pushers with very few changes to the pushers.

This commit also adds a couple tests and improves a couple old tests. Which should provide a guarantee that it actually works. The tests include data continuity tests (ensure that a timeseries consisting of natural numbers is continuous, even after the pusher fails for a bit), which is a pretty strong test.