Initial commit for backfill rewrite
This commit adds backfill. It required a complete rewrite of how
historyRead is done in UAClient, moving the main responsibility for
continuation out of UAClient (A good thing, UAClient is a bit bloated
and hard to test), into a new HistoryReader class.

There are no tests for backfill yet, though all the existing tests
running frontfill only still pass, which is a good sign.

The new feature is a bit limited, in that there is no local state, so
backfill will have to run each startup. Unlike PI this isn't a very big
problem, as all it does is effectively double the number of initial
requests, though of course it could be better. The backfilling and
frontfilling are in practice pretty much the same. Backfill runs after
frontfill, backwards from the earliest known timestamp.

This should be useful in cases where there is a lot of history, as it
should allow us to start streaming earlier. If there isn't much history,
it can be disabled to just use the old behavior, with no performance
cost.

Required some changes to config, most have been added to a new History
category, and removed from various others: HistoryGranularity,
HistoryNodesChunk HistoryChunk, History and HistoryStartTime have been
moved from Source and Extraction to History, EarliestChunk has been
added to CDFPusher config. The latter is not breaking.