Add tests for data to backfill
Backfill on data now works, events is more complicated, as I'm still not
entirely sure if we want to do it this way, relying on the timestamp of
the event when extracting (this will mean that events triggered in the
past while the extractor is offline will not be detected), or try for
some other way. Issue is that this "other way" may not be supported, and
python-opcua fetches events directly on triggered time, I'm suspecting
this is a common interpretation.

The alternative is to essentially get all events, then filter later, but
this is very expensive, and not very clean.