More tests, implement system for getting first/last event
This is part of a future plan to create a local database to store state
between each run. There is a new config option to all pushers for
disabling getting first/last timestamps for both events and data. This
means that we can have a "local state" pusher, which fits neatly into
what we already have. The idea has always been that pushers can
implement a subset of available features and still work as expected.

The tests are mostly for influxdb, a simple test for events and a couple
for data. Tests are still needed to test restarting influxdb with
events, and restarting if both influxdb and cdf pushers are present.