Test restart when backfill is enabled with influxdb
The purpose of these tests is mainly just to test that the first/last logic
works properly, that backfill starts from somewhere reasonable after
restart.

This commit also removes the old latestEvent.bin file, and everything
related to that. It really wasn't very compatible with the new backfill
logic, and it should really be replaced if we want to properly support
historical events.