Backfill (#91)
Implements backfill of data and events. Unlike PI this is almost exactly the same as frontfill in practice: HistoryRead in OPC-UA takes start/end timestamps, if the order is reversed (start is after end) then the data is read backwards. Thus there is actually a single method that does the history read itself.

In order to make this cleaner, I moved part of the logic into its own class, instead of doing things using callbacks like before. In practice there is little difference, but the code is a bit cleaner. The "state" of the backfill/frontfill is kept with a params object that contains the state of each node being read.

Of course, since this is OPC-UA it needs to be possible to extract history using just frontfill, as there is no guarantee that servers will support backwards reading (although that is a breach of the specification, it seems quite a lot of servers are lacking in that regard). By default backfill is disabled, and all the tests still pass with no changes, so this is an optional feature.

Some major changes were however made to config, in order to make that cleaner as well. History read parameters were spread in a fairly nonsensical way all over the different categories. Now they are gathered in a single config block.

The most complex part of this change is reading first/last timestamps. Part of this is also adding the ability to read first/last event. No attempt was yet made to do that in CDF, though it is likely possible, it might be expensive, especially the last event might require multiple requests to a huge number of different nodes. This is however trivial in influxdb, where events are represented as string timeseries.

Thus a solution is to run both CDF and influxdb, in which case events will be read solely based on influxdb. If other pushers are added later they might also be able to read first/last timestamps.

In the long term, it might be relevant to add a local database storage to the CDFPusher, or as its own pusher, in order to buffer timestamps, both for performance and for backfill of events. For the time being it is not that urgent, however.