Refactor launch into a separate project.
Both Extractor and ConfigurationTool are now netstandard2.1 apps (not
that this makes much of a difference). A separate ExtractorLauncher
project serves as the execution root.

To make setup easier ExtractorLauncher supports a couple of command-line
arguments for a basic setup, and can also launch the config-tool. The
consequence of this is that it is possible to never have to edit any
config-files manually, provided the results from the config-tool are
good enough.