Feature/configuration tool (#93)
Separate tool that uses UAClient to map out server characteristics. The idea is to avoid the large number of issues we get with customers due to the great variety of OPC-UA implementations. Experience has shown that a great many of these are lacking in terms of features, and often flawed. We have no choice but to work around this, and this tool should make that easier.

It also makes it easier to configure the extractor, reducing the amount of work that needs to be done looking through uaExpert.

As part of this, in order to make setup easier, launching has been moved to a separate ExtractorLauncher project, which can launch both the extractor and the config tool, based on command line arguments. For convenience, a couple command-line options have been added which allow us to run the config-tool without any config file, which can then generate a config file for the extractor.

The intended flow of setup can now be: Launch the config tool with no config file -> Use the generated config file to run the extractor directly.

On a functional level the separation of the launch from code should cause no difference for users. An important difference to note is that both the Extractor and the ConfigurationTool are built as libraries instead of console-apps, which causes some differences is usage code-wise.