Fixed ugly bug in influxpusher
Apparently, both inconsistent tags and missing tag values cause issues
with influxdb. This is a problem when dealing with events, so in order
to fix this multiple things had to be done:
  - Separate each event type into its own measurement, so that
  measurementnames are on the form
  events.[globalprefix][sourcenodeid][eventtypeid]
  - Set empty fields to string "null".

In order to improve the code a bit (suddenly we had a very unpleasant
number of potential measurements being read), I changed it so that when
reading first/last timestamp for events (not datapoints yet, it is
probably a waste of time to do it there, for the most part), we first
list all measurements, then fetch only from relevant measurements.