1.1 Fixes (#94)
PR for misc fixes ahead of the 1.1 release.

Update CogniteSdk

Add fxCop analyzers and fix or disable warnings related to this. The disabled warnings are generally things that aren't that important and are frequently violated in this project. Otherwise they are disabled locally. The fixed warnings are mostly small things that may prevent or give clearer errors in the future, but there is also quite a few cases of disposable objects not being properly disposed, which may have created memory leaks.

Fix ForceRestart option, add a new Source/ExitOnFailure option, which prevents the extractor from restarting internally on error, instead deferring to some system specific scheduler.

Add tests for reconnecting to a server with loss of connection. Due to the continuity tests we do, stopping the server isn't ideal, and doesn't really give us any extra information here. Instead we use ncat to create a temporary proxy from 4839 to 4840, connect to that, then kill the proxy to break the connection.

Add metrics configuration like PI, with multiple pushers and support for a server for scraping.

Add context information to logs, like PI.

Add metric for version information.

Cleaned up exceptions and improved error reporting.

Improved results from configtool, to be clearer.

Fixed major issue in InfluxPusher, related to how events are pushed. This changes the dynamic of events in influxdb in two ways:
 - First, events are now stored in different measurements depending on type, so we get [globalprefix][sourcenode]:[eventtype].
 - Second, null values are now represented with "null", since influxdb line protocol does not support missing tags.