Add and improve a few tests (#97)
Add test for failure to connect with a critical pusher.

Add a few custom data types and some built in (Range and EUInformation) to the array server, and also test these.

It turns out custom data types were broken (default behavior was treating them as numeric). This fixes that and also adds tests to verify that it works as expected.

It is worth noting that custom datatypes is somewhat buggy in python-opcua, though not so bad that it's impossible to write tests for it. Notably, the new datatypes are not really treated as datatypes but as ObjectTypes, which creates some issues for the config tool, which filters on datatype. It shouldn't really do any harm to also allow for ObjectTypes in the DataType hierarchy, however.