Feature/litedb storage (#100)
Huge patch that does a couple of different, related things:

Persist state in a local influxdb database periodically, then use this as a source for extracted ranges on startup. The idea is that you would disable range reading on other pushers and only use this, which would be generally sufficient. The old system of reading ranges from CDF/influxdb remains, and can be turned on to re-read parts of history.

This involved a major refactoring of the data-flow in the extractor. Before, each pusher had its own queues for datapoints and events, which allowed them to push at individual rates. This isn't really viable anymore, due to this and new structure in FailureBuffer. Now, there is a central queue in Extractor, and all pushers push simultaneously and wait for each other. This costs a little flexibility, but it likely helps in terms of performance, and makes this patch possible.

Remove the binary buffer file and replace it with LiteQueue; A FIFO queue built on LiteDB. This works very well, as it has transaction support, so we can write to it on failure, then read from it once we re-connect to the pushers, but only commit once we actually succeed in pushing datapoints/events. It also allows buffering events, which was never completed before.

Add state storage for influxdb buffering. The influxdb buffer remains, because in some cases it still makes sense, however in order to make it as reliable as the queue, I've added a system to persist the ranges of the failed points to litedb.