Feature/old buffer (#101)
Re-add the old system of buffering to a file, but expanded to cover both events and string datapoints, as well as handle larger amounts of data.

It turns out buffering with litequeue is too slow to handle large amounts of data, the old system is, and this solution should be a bit more stable than the old one.

The general idea is the same as before, For example, the format of string datapoints is

`[ushort][ushort][string][double][timestamp][isstring][ushort][string]`

First the length of the entire object, then before each string, the length of the string. A similar system is used for events.

Thinks will break badly if we ever get an event of more than 64kB. But CDF events are limited to about 12kB at the most, so it shouldn't really be an issue.