No longer abort startup if pushers fail
This means that the extractor should never crash due to failure of a
pusher. Instead, wait with initialization until a connection test
succeeds, then initialize late as a part of the pusher loop.

Startup is starting to get very complex, which is scary, but this
feature should work, all in all it is not very complicated.

In order to do this properly it was necessary to move the connection
test logic into the main pusher loop, and that results in fewer requests
to CDF to begin with, which is a good thing in general.