Feature/defer startup (#103)
A relatively small change to the extractor to allow pushing of nodes to fail on startup. If this happens, start pushing as usual, and do initialization once a connection test succeeds.

Part of this means that the connection test succeeding is important for performance. In general the system is designed so that even if the connection test gives a false positive, we won't lose consistency, but it will still mean a hit to performance.

This patch reflects a small change in design philosophy. The pushers should no longer be able to greatly affect the extractor, meaning that in general extraction goes on as usual even if the pushers don't work at all. The exception is the state storage: If one of the pushers don't receive data, we won't update saved states, but this is as expected. Perhaps in the future we could make it so that a pusher can be marked as non-essential, allowing state storage to happen even if a pusher is offline. That is edge-case, however. If there are many pushers it might be better to not use state storage, and instead initialize states using pushers. In this case, they generally operate independently.

History will not be run if there are no available pushers on startup, but since pushing starts, any enabled system for buffering datapoints will work they would even if the pusher had stopped working after initialization.

Remove the Critical option, since it doesn't make sense anymore, and it was never really that great.