Feature/mqtt pusher (#110)
Add MQTT Pusher and new MQTTCDFBridge project.

The goal of this is to create a new pipeline to CDF, without any communication from CDF to the extractor. Fortunately, the general extractor design supports this nicely (each pusher may implement only a subset of possible features, the rest are covered by other pushers or litedb), so the only changes to the extractor was the addition of a new pusher.

The issues occur when posting this to CDF, as a CDF request with 1000 assets will fail if there is a single duplicate. To fix this, it is necessary to create a service that parses MQTT messages, tests them against CDF, and does whatever else is necessary to transform them to a final state. This involves removing duplicates, removing invalid parentIds of timeseries and events, removing wrong datapoint types, and transforming externalIds to interalIds for timeseries and events.

The MQTT-pusher is designed to be completely independent from OPC-UA. It takes arrays of normal objects from the SDK, except for timeseries and events, where it is possible to augment them a little with "assetExternalId(s)", which are translated to interalIds.

Tests have been added in two blocks: standalone tests of the Bridge and tests of the MQTTPusher. If need be the standalone tests can be separated out of the project entirely without much effort. To this end the standalone tests do not use the tester. The extractor-side testing is fairly basic, but a lot of behavior is common, and the pusher is well covered. Failure of the broker is not really something that is part of a common use-case.