Hopefully fix tests. (#112)
The primary issue seems to have been enabled paralellization in the MQTT tests. Since all tests use the same client configuration, they also share the same client id and certificate, meaning that server-side client state is shared. This can cause very unpredictable issues, as publish messages randomly only work for one of the active clients, one calling close causes the others to lose connection, etc.

It would be interesting to see if there is some mechanism to check for already connected clients. This might have caused issues in production as well, if a process is left behind somehow.

There are still some issues with the cdfbackfillevents test. I managed to reproduce it briefly, added some logging, then it disappeared again. If it fails in the future I might be able to fix it.