Add test-server implementation
This is heavily based on samples from the UA.NET repository. It
currently implements the address space for each server, as well as
features necessary for each of the different server implementations
(historical data access, historical event access, proof of concept for
auditing).

The idea is to use this to increase the level of control we have over
the server during testing. This will allow us to re-initialize the
server before each test if desired, to make sure each test operates on a
clean server.

It also gives us much greater control of when we update nodes, trigger
events, etc. It may be possible to avoid all periodic updates entirely.

Finally it also allows us to reference the server contents. So instead
of testing for some sort of continuity or similar, we can actually test
for corresponendence between the contents of the server and the
extractor, which is very powerful.

The SDK server implementation does have some limitations, fortunately
none of the affect the current extractor implementation. If future
features of the extractor requires these features (these are typically
more complex forms of reading data, or features to edit the server from
the client), the server appears to be built in such a way that it is
possible for us to add these features by extending the server object.