Update tests, add queue and sampling to config
There is a clear difference between sampling-rate and publishing
interval. This should be reflected in the config file. Default values
are probably fine for the most part, but there is a chance that points
are skipped if these are wrong.

This is unfortunately a breaking change, though I don't think this
particular config option has seen much use.