Modify influxdb tests
This required some further changes to the server implementation. The
issue was with historizing events on the server node. CustomNodeManager2
is not built to handle history on nodes it doesn't manage directly. If
there is some other way to do this, I couldn't find it. In the end I
just made an exception especially for the server node. Fortunately, the
higher level node managers are capable of reading history from any child
node. I might also have been able to handle this by modifying
GetManagerHandle (look into this?)