Add some tests to events, fix bug
Fixed an old bug in influxPusher. It took reliable, predictable testing
to discover it. The issue was that the query I used was /...i=2*/,
meaning ...i= followed by one or more 2's, so it matched wrong hits, and
this meant that if the bad areas overlapped at all the first hit on a
series would overwrite later ones, resulting in very unpredictably
missed events. This has probably been causing issues for a long time.

The fix is to simply change it to /...i=2:.*/, ...i=2: followed by
anything.