Feature/test server (#114)
Adding features to the extractor has become increasingly difficult due to the unpredictability of testing against the python servers. The unpredictable timing and ordering of tests means that a majority of test runs ends with some error that can not be reproduced. The Python library also takes some liberties in making assumptions about settings, which can become a pitfall if we want to connect to servers which do not make this assumption in the future.

In order to fix this we have to move away from the python servers entirely, and create a new server implementation in the .NET UA SDK. This is much more complicated than doing it in python, since the server SDK leaves a large part of the implementation to the user, and is not documented at all. There are a number of sample servers here, which this implementation is based on.

This patch reproduces the old server configurations in the .NET SDK, but with several major advantages:

- The server can be completely reset between each test, ensuring that problems are not caused by lack of cleanup on the server.
- The server can be populated with a fixed number of historical events or datapoints before starting, and new points can be generated on demand rather than on a fixed interval, this greatly increases the predictability of tests.
- The tests can use server data as a reference.
- The development even uncovered a few subtle issues which the python server covered up, or which vanished due to unpredictable behavior. This patch will hopefully make development easier in the future, and make it possible to improve the existing tests further.

It makes a few changes to the main config schema, in response to a few issues uncovered:

- PollingInterval has been split into PublishingInterval and SamplingInterval, which reflects the actual behavior, and uses terms from the OPC-UA standard.
- History.Enabled now affects events. This was just unintuitive before. If I set History to disabled, then I expect history to be disabled.
- In order to allow disabling datapoint history while keeping event history enabled, a new History.Data option has been created, it is true by default, so this change is backwards compatible, except in the case of historical events (I don't believe that has ever been used in production, however).