Flexible behavior when ValueRank is wrong (#123)
A small change to make ValueRank more flexible. This is one of those things that people writing OPC-UA interfaces for their applications don't always handle in a way that works for us.

Essentially, since we map each variable to a number of timeseries, we require that the variables are of fixed size. In OPC-UA this is indicated by setting the ValueRank to Scalar, or to something else and setting ArrayDimensions to some value.

If somebody decides to not do this, we have ignored them. This can cause issues due to laziness on the server side: For example, the OPC-UA server maps some OT-system, where the values can be both scalar and arrays. The engineers working with the OT-system knows that all the raw values are scalar, but the server software lacks the ability to specify that, or it is difficult to do, so they just set them all as "Any", or "ScalarOrOneDimension", and of course do not specify ArrayDimensions.

The extractor then ignores them all.

This PR introduces general flexibility, in that if a variable considered scalar in the extractor returns an array, we log a verbose error and increment a metric, then just pick the first element. We do something similar if we get a longer/shorter array than ArrayDimension specify (this violates the specification, but historically a lot of developers don't care).

It also adds a config option to treat all "unspecific" ValueRanks as scalar, which is generally better than nothing, though not ideal.