Read data for nodes on rebrowse if needed
A bit of a quick and dirty way to do it, by just rediscovering all nodes
when browsing. Only modified nodes are passed to pushers.

In practice, if we want to update we need to read data from nodes. The
only case where I think this could have been more efficient is if we
only wanted to update certain nodes or similar, but it probably isn't
worth the effort. This does a full re-read, effectively restarting the
extractor without doing any synchronization, and without losing active
subscriptions or creating any holes in extraction.

Changes in array dimensions or datatype have not been covered yet, they
are considerably more difficult to do anything about, and pretty
edge-case. That will not be a part of this feature.