Feature/raw metadata (#124)
Push metadata/context to raw instead of to assets/timeseries. Timeseries are pushed with just minimal data (isString, isStep and externalId), the full asset and timeseries objects are pushed directly to raw. By far the simplest solution here was just to directly transform asset/timeseries objects from the SDK into json and push data, if that means we get some extra fields we don't need in raw, that should be largely irrelevant.

Led to some changes in the extractor:

Now that since assets are not created, events have to be created without associated assets, but this isn't really a bad thing. In fact, it makes sense to do so either way. Events now largely ignore the SourceNode property, and can have null SourceNode. It is still used for context in CDF if it exists, and it is added to metadata, but it is no longer required, and no longer filtered on the server.

This was overdue, having too large Where-clauses in the subscription can cause issues for the server, and was always a bit too complicated. If we want to, we can add filtering in the extractor at a later date.

This created issues in the influxdb pusher, which used SourceNode for measurement names of events. This was a stupid choice to begin with, and has been changed to using the emitter.