Feature/updates (#126)
Update assets and timeseries in CDF and CDF Raw.

Adds new config to enable updating assets and timeseries in CDF. Each different thing that can be updated is separately configured:
 - name, DisplayName in OPC-UA
 - description
 - metadata, which requires re-reading properties in OPC-UA
 - context, meaning assetId/parentExternalId.
Updates are done carefully. Metadata is not removed, only added or overwritten. If description/name/context is set to null in OPC-UA, it is not removed from CDF.

Changes are discovered on restart or on re-browse. Re-browse period can be configured (this is not a new feature). If updates are not enabled there should be no additional load on the server, if they are, and especially if metadata is enabled, there will be some additional load on startup or re-browse.

Changes in ArrayDimensions are not handled. This is difficult to do anything about, not something we want to have happen, and very edge-case.

This PR got pretty big. Testing accounts for about half, since this required expanding mocking to support updates, and writing code for making various changes to the server.

Both raw and using updates, MQTT will come at a later date.

* Better structure when updating raw

* Add test

* Test for raw updates

* Add test for changing property name

Since we don't remove it should add a new metadata field.

* Example config

* Fix failing test

* Consolidate common code in extractor

* Update packages, fix parentExternalId issue